REFACTOR: reorganize folder structure
