+++ src/model/Robot.ts
@@ -1,5 +1,5 @@
 import Joint from '@/model/Joint'
-import Tool from '@/model/Tool'
+import Tool from '@/model/Tools/Tool'
 
 export default class Robot<T extends Tool<any>> {
   constructor(public joints: Array<Joint>, public tool: T) {}
